requirements.txt
Django~=6.0
djangorestframework~=3.16.1
python-dotenv~=1.2.1
asgiref~=3.11.0
certifi~=2025.11.12
charset-normalizer~=3.4.4
idna~=3.11
pip~=25.3
psycopg2-binary~=2.9.11
requests~=2.32.5
sqlparse~=0.5.5
tzdata~=2025.3
urllib3~=2.6.2
dj_database_url~=3.1.0
gunicorn~=23.0.0
django-cors-headers~=4.9.0
drf-spectacular~=0.29.0